+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/tools/ExplosiveShovel.java
@@ -1,28 +1,18 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.tools;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import org.bukkit.Bukkit;
 import org.bukkit.Effect;
 import org.bukkit.Material;
-import org.bukkit.Sound;
 import org.bukkit.block.Block;
-import org.bukkit.event.block.BlockBreakEvent;
-import org.bukkit.event.block.BlockExplodeEvent;
+import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.cscorelib2.materials.MaterialTools;
 import io.github.thebusybiscuit.cscorelib2.protection.ProtectableAction;
-import io.github.thebusybiscuit.slimefun4.api.items.ItemSetting;
-import io.github.thebusybiscuit.slimefun4.core.attributes.DamageableItem;
-import io.github.thebusybiscuit.slimefun4.core.attributes.NotPlaceable;
 import me.mrCookieSlime.Slimefun.SlimefunPlugin;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SimpleSlimefunItem;
-import me.mrCookieSlime.Slimefun.Objects.handlers.BlockBreakHandler;
-import me.mrCookieSlime.Slimefun.api.Slimefun;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
 /**
@@ -34,77 +24,28 @@ import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
  * @see ExplosivePickaxe
  *
  */
-public class ExplosiveShovel extends SimpleSlimefunItem<BlockBreakHandler> implements NotPlaceable, DamageableItem {
-
-    private final ItemSetting<Boolean> damageOnUse = new ItemSetting<>("damage-on-use", true);
+public class ExplosiveShovel extends ExplosiveTool {
 
     public ExplosiveShovel(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
-
-        addItemSetting(damageOnUse);
     }
 
     @Override
-    public BlockBreakHandler getItemHandler() {
-        return new BlockBreakHandler() {
+    protected void breakBlock(Player p, ItemStack item, Block b, int fortune, List<ItemStack> drops) {
+        if (MaterialTools.getBreakableByShovel().contains(b.getType()) && SlimefunPlugin.getProtectionManager().hasPermission(p, b.getLocation(), ProtectableAction.BREAK_BLOCK)) {
+            SlimefunPlugin.getProtectionManager().logAction(p, b, ProtectableAction.BREAK_BLOCK);
 
-            @Override
-            public boolean isPrivate() {
-                return false;
-            }
+            b.getWorld().playEffect(b.getLocation(), Effect.STEP_SOUND, b.getType());
 
-            @Override
-            public boolean onBlockBreak(BlockBreakEvent e, ItemStack item, int fortune, List<ItemStack> drops) {
-                if (isItem(item)) {
-                    if (Slimefun.hasUnlocked(e.getPlayer(), ExplosiveShovel.this, true)) {
-                        e.getBlock().getWorld().createExplosion(e.getBlock().getLocation(), 0.0F);
-                        e.getBlock().getWorld().playSound(e.getBlock().getLocation(), Sound.ENTITY_GENERIC_EXPLODE, 0.3F, 1F);
-
-                        List<Block> blocks = new ArrayList<>();
-                        for (int x = -1; x <= 1; x++) {
-                            for (int y = -1; y <= 1; y++) {
-                                for (int z = -1; z <= 1; z++) {
-                                    if (x == 0 && y == 0 && z == 0) {
-                                        continue;
-                                    }
-
-                                    blocks.add(e.getBlock().getRelative(x, y, z));
-                                }
-                            }
-                        }
-
-                        BlockExplodeEvent blockExplodeEvent = new BlockExplodeEvent(e.getBlock(), blocks, 0);
-                        Bukkit.getServer().getPluginManager().callEvent(blockExplodeEvent);
-                        if (!blockExplodeEvent.isCancelled()) {
-                            blockExplodeEvent.blockList().forEach(b -> {
-                                if (MaterialTools.getBreakableByShovel().contains(b.getType()) && SlimefunPlugin.getProtectionManager().hasPermission(e.getPlayer(), b.getLocation(), ProtectableAction.BREAK_BLOCK)) {
-                                    SlimefunPlugin.getProtectionManager().logAction(e.getPlayer(), b, ProtectableAction.BREAK_BLOCK);
-
-                                    b.getWorld().playEffect(b.getLocation(), Effect.STEP_SOUND, b.getType());
-
-                                    for (ItemStack drop : b.getDrops(getItem())) {
-                                        if (drop != null) {
-                                            b.getWorld().dropItemNaturally(b.getLocation(), drop);
-                                        }
-                                    }
-
-                                    b.setType(Material.AIR);
-                                    damageItem(e.getPlayer(), item);
-                                }
-                            });
-                        }
-                    }
-
-                    return true;
+            for (ItemStack drop : b.getDrops(getItem())) {
+                if (drop != null) {
+                    b.getWorld().dropItemNaturally(b.getLocation(), drop);
                 }
-                else return false;
             }
-        };
-    }
 
-    @Override
-    public boolean isDamageable() {
-        return damageOnUse.getValue();
+            b.setType(Material.AIR);
+            damageItem(p, item);
+        }
     }
 
 }